Broke TigerHelpers up into multiple files
