+++ src/main/java/com/titaniumtigers4829/TigerHelpers.java
@@ -2,24 +2,15 @@
 
 package com.titaniumtigers4829;
 
-import edu.wpi.first.math.geometry.Pose2d;
-import edu.wpi.first.math.geometry.Pose3d;
-import edu.wpi.first.math.geometry.Rotation2d;
-import edu.wpi.first.math.geometry.Rotation3d;
-import edu.wpi.first.math.geometry.Translation2d;
-import edu.wpi.first.math.geometry.Translation3d;
-import edu.wpi.first.math.util.Units;
-import edu.wpi.first.networktables.DoubleArrayEntry;
-import edu.wpi.first.networktables.NetworkTable;
-import edu.wpi.first.networktables.NetworkTableEntry;
-import edu.wpi.first.networktables.NetworkTableInstance;
-import edu.wpi.first.networktables.TimestampedDoubleArray;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
 import com.titaniumtigers4829.data.IMUData;
 import com.titaniumtigers4829.data.PoseEstimate;
+import com.titaniumtigers4829.data.PoseEstimate.Botpose;
 import com.titaniumtigers4829.data.RawFiducial;
+import com.titaniumtigers4829.utils.DataUtils;
+import com.titaniumtigers4829.utils.NTUtils;
+import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.geometry.Pose3d;
+import edu.wpi.first.networktables.NetworkTableEntry;
 
 /**
  * TigerHelpers is built on top of LimelightHelpers, providing a set of static methods and classes
@@ -31,309 +22,6 @@ import com.titaniumtigers4829.data.RawFiducial;
  */
 public class TigerHelpers {
 
-  private static final Map<String, DoubleArrayEntry> doubleArrayEntries = new ConcurrentHashMap<>();
-
-  private static final String sanitizeName(String name) {
-    if (name.equals("") || name.equals(null)) {
-      return "limelight";
-    }
-    return name;
-  }
-
-  /**
-   * Takes a 6-length array of pose data and converts it to a Pose3d object. Array format: [x, y, z,
-   * roll, pitch, yaw] where angles are in degrees.
-   *
-   * @param inData Array containing pose data [x, y, z, roll, pitch, yaw]
-   * @return Pose3d object representing the pose, or empty Pose3d if invalid data
-   */
-  public static Pose3d toPose3D(double[] inData) {
-    if (inData.length < 6) {
-      return new Pose3d();
-    }
-    return new Pose3d(
-        new Translation3d(inData[0], inData[1], inData[2]),
-        new Rotation3d(
-            Units.degreesToRadians(inData[3]),
-            Units.degreesToRadians(inData[4]),
-            Units.degreesToRadians(inData[5])));
-  }
-
-  /**
-   * Takes a 6-length array of pose data and converts it to a Pose2d object. Uses only x, y, and yaw
-   * components, ignoring z, roll, and pitch. Array format: [x, y, z, roll, pitch, yaw] where angles
-   * are in degrees.
-   *
-   * @param inData Array containing pose data [x, y, z, roll, pitch, yaw]
-   * @return Pose2d object representing the pose, or empty Pose2d if invalid data
-   */
-  public static Pose2d toPose2D(double[] inData) {
-    if (inData.length < 6) {
-      return new Pose2d();
-    }
-    Translation2d tran2d = new Translation2d(inData[0], inData[1]);
-    Rotation2d r2d = new Rotation2d(Units.degreesToRadians(inData[5]));
-    return new Pose2d(tran2d, r2d);
-  }
-
-  /**
-   * Converts a Pose3d object to an array of doubles in the format [x, y, z, roll, pitch, yaw].
-   * Translation components are in meters, rotation components are in degrees.
-   *
-   * @param pose The Pose3d object to convert
-   * @return A 6-element array containing [x, y, z, roll, pitch, yaw]
-   */
-  public static double[] pose3dToArray(Pose3d pose) {
-    double[] result = new double[6];
-    result[0] = pose.getTranslation().getX();
-    result[1] = pose.getTranslation().getY();
-    result[2] = pose.getTranslation().getZ();
-    result[3] = Units.radiansToDegrees(pose.getRotation().getX());
-    result[4] = Units.radiansToDegrees(pose.getRotation().getY());
-    result[5] = Units.radiansToDegrees(pose.getRotation().getZ());
-    return result;
-  }
-
-  /**
-   * Converts a Pose2d object to an array of doubles in the format [x, y, z, roll, pitch, yaw].
-   * Translation components are in meters, rotation components are in degrees. Note: z, roll, and
-   * pitch will be 0 since Pose2d only contains x, y, and yaw.
-   *
-   * @param pose The Pose2d object to convert
-   * @return A 6-element array containing [x, y, 0, 0, 0, yaw]
-   */
-  public static double[] pose2dToArray(Pose2d pose) {
-    double[] result = new double[6];
-    result[0] = pose.getTranslation().getX();
-    result[1] = pose.getTranslation().getY();
-    result[2] = 0;
-    result[3] = Units.radiansToDegrees(0);
-    result[4] = Units.radiansToDegrees(0);
-    result[5] = Units.radiansToDegrees(pose.getRotation().getRadians());
-    return result;
-  }
-
-  private static double extractArrayEntry(double[] inData, int position) {
-    if (inData.length < position + 1) {
-      return 0;
-    }
-    return inData[position];
-  }
-
-  private static PoseEstimate unpackBotPoseEstimate(
-      String limelightName, String entryName, boolean isMegaTag2) {
-    DoubleArrayEntry poseEntry =
-        TigerHelpers.getLimelightDoubleArrayEntry(limelightName, entryName);
-
-    TimestampedDoubleArray tsValue = poseEntry.getAtomic();
-    double[] poseArray = tsValue.value;
-    long timestamp = tsValue.timestamp;
-
-    if (poseArray.length == 0) {
-      // Handle the case where no data is available
-      return new PoseEstimate();
-    }
-
-    Pose2d pose = toPose2D(poseArray);
-    double latency = extractArrayEntry(poseArray, 6);
-    int tagCount = (int) extractArrayEntry(poseArray, 7);
-    double tagSpan = extractArrayEntry(poseArray, 8);
-    double tagDist = extractArrayEntry(poseArray, 9);
-    double tagArea = extractArrayEntry(poseArray, 10);
-
-    // Convert server timestamp from microseconds to seconds and adjust for latency
-    double adjustedTimestamp = (timestamp / 1000000.0) - (latency / 1000.0);
-
-    RawFiducial[] rawFiducials = new RawFiducial[tagCount];
-    int valsPerFiducial = 7;
-    int expectedTotalVals = 11 + valsPerFiducial * tagCount;
-
-    if (poseArray.length != expectedTotalVals) {
-      // Don't populate fiducials
-    } else {
-      for (int i = 0; i < tagCount; i++) {
-        int baseIndex = 11 + (i * valsPerFiducial);
-        int id = (int) poseArray[baseIndex];
-        double txnc = poseArray[baseIndex + 1];
-        double tync = poseArray[baseIndex + 2];
-        double ta = poseArray[baseIndex + 3];
-        double distToCamera = poseArray[baseIndex + 4];
-        double distToRobot = poseArray[baseIndex + 5];
-        double ambiguity = poseArray[baseIndex + 6];
-        rawFiducials[i] = new RawFiducial(id, txnc, tync, ta, distToCamera, distToRobot, ambiguity);
-      }
-    }
-
-    return new PoseEstimate(
-        pose,
-        adjustedTimestamp,
-        latency,
-        tagCount,
-        tagSpan,
-        tagDist,
-        tagArea,
-        rawFiducials,
-        isMegaTag2);
-  }
-
-  /**
-   * Gets the latest raw fiducial/AprilTag detection results from NetworkTables.
-   *
-   * @param limelightName The name of the Limelight set in the UI ("" for default)
-   * @return Array of RawFiducial objects containing detection details
-   */
-  public static RawFiducial[] getRawFiducials(String limelightName) {
-    NetworkTableEntry entry =
-        TigerHelpers.getLimelightNetworkTableEntry(limelightName, "rawfiducials");
-    double[] rawFiducialArray = entry.getDoubleArray(new double[0]);
-    int valsPerEntry = 7;
-    if (rawFiducialArray.length % valsPerEntry != 0) {
-      return new RawFiducial[0];
-    }
-
-    int numFiducials = rawFiducialArray.length / valsPerEntry;
-    RawFiducial[] rawFiducials = new RawFiducial[numFiducials];
-
-    for (int i = 0; i < numFiducials; i++) {
-      int baseIndex = i * valsPerEntry;
-      int id = (int) extractArrayEntry(rawFiducialArray, baseIndex);
-      double txnc = extractArrayEntry(rawFiducialArray, baseIndex + 1);
-      double tync = extractArrayEntry(rawFiducialArray, baseIndex + 2);
-      double ta = extractArrayEntry(rawFiducialArray, baseIndex + 3);
-      double distToCamera = extractArrayEntry(rawFiducialArray, baseIndex + 4);
-      double distToRobot = extractArrayEntry(rawFiducialArray, baseIndex + 5);
-      double ambiguity = extractArrayEntry(rawFiducialArray, baseIndex + 6);
-
-      rawFiducials[i] = new RawFiducial(id, txnc, tync, ta, distToCamera, distToRobot, ambiguity);
-    }
-
-    return rawFiducials;
-  }
-
-  private static NetworkTableInstance networkTableInstance = NetworkTableInstance.getDefault();
-
-  /**
-   * Sets the NetworkTableInstance to use for Limelight communication. Normally it it uses the
-   * default instance. This is useful for unit testing when you might need multiple instances.
-   */
-  public static void setNetworkTableInstance(NetworkTableInstance networkTableInstance) {
-    TigerHelpers.networkTableInstance = networkTableInstance;
-    clearCaches();
-    flushNetworkTable();
-  }
-
-  /**
-   * Gets the NetworkTableInstance used for Limelight communication.
-   *
-   * @return The NetworkTableInstance used for Limelight communication
-   */
-  public static NetworkTableInstance getNetworkTableInstance() {
-    return networkTableInstance;
-  }
-
-  /**
-   * Clears any static caches held in TigerHelpers. This should be called in your test setup (or
-   * teardown) to ensure a fresh state between tests.
-   */
-  public static void clearCaches() {
-    doubleArrayEntries.clear();
-  }
-
-  /**
-   * Flushes the NetworkTables to ensure that any newly updated entry is immediately sent to the
-   * network.
-   */
-  public static void flushNetworkTable() {
-    networkTableInstance.flush();
-  }
-
-  /**
-   * Gets the NetworkTable for a Limelight camera.
-   *
-   * @param tableName The name of the Limelight set in the UI ("" for default)
-   * @return NetworkTable for the Limelight
-   */
-  public static NetworkTable getLimelightNetworkTable(String tableName) {
-    return networkTableInstance.getTable(sanitizeName(tableName));
-  }
-
-  /**
-   * Gets a NetworkTableEntry from a Limelight camera.
-   *
-   * @param tableName The name of the Limelight set in the UI ("" for default)
-   * @param entryName Name of the entry to get
-   * @return NetworkTableEntry for the entry name
-   */
-  public static NetworkTableEntry getLimelightNetworkTableEntry(
-      String tableName, String entryName) {
-    return getLimelightNetworkTable(tableName).getEntry(entryName);
-  }
-
-  /**
-   * Gets a DoubleArrayEntry from a Limelight camera. This is useful for getting raw pose data.
-   *
-   * @param tableName The name of the Limelight set in the UI ("" for default)
-   * @param entryName Name of the entry to get
-   * @return DoubleArrayEntry for the entry name
-   */
-  public static DoubleArrayEntry getLimelightDoubleArrayEntry(String tableName, String entryName) {
-    String key = tableName + "/" + entryName;
-    return doubleArrayEntries.computeIfAbsent(
-        key,
-        k -> {
-          NetworkTable table = getLimelightNetworkTable(tableName);
-          return table.getDoubleArrayTopic(entryName).getEntry(new double[0]);
-        });
-  }
-
-  /**
-   * Gets a NetworkTableEntry from a Limelight camera as a double.
-   *
-   * @param tableName The name of the Limelight set in the UI ("" for default)
-   * @param entryName Name of the entry to get
-   * @return double value of the entry
-   */
-  public static double getLimelightNetworkTableDouble(String tableName, String entryName) {
-    return getLimelightNetworkTableEntry(tableName, entryName).getDouble(0.0);
-  }
-
-  /**
-   * Sets a NetworkTableEntry from a Limelight camera as a double.
-   *
-   * @param tableName The name of the Limelight set in the UI ("" for default)
-   * @param entryName Name of the entry to set
-   * @param val Value to set the entry to
-   */
-  public static void setLimelightNetworkTableDouble(
-      String tableName, String entryName, double val) {
-    getLimelightNetworkTableEntry(tableName, entryName).setDouble(val);
-  }
-
-  /**
-   * Sets a NetworkTableEntry from a Limelight camera as a double array. If you want to the network
-   * table array for a {@link PoseEstimate}, use {@link #setBotPoseEstimate(PoseEstimate, String)},
-   * if for the {@link RawFiducial}s use {@link #setRawFiducials(RawFiducial[], String)}.
-   *
-   * @param tableName The name of the Limelight set in the UI ("" for default)
-   * @param entryName Name of the entry to set
-   * @param val Value to set the entry to
-   */
-  public static void setLimelightNetworkTableDoubleArray(
-      String tableName, String entryName, double[] val) {
-    getLimelightNetworkTableEntry(tableName, entryName).setDoubleArray(val);
-  }
-
-  /**
-   * Gets a NetworkTableEntry from a Limelight camera as a double array.
-   *
-   * @param tableName The name of the Limelight set in the UI ("" for default)
-   * @param entryName Name of the entry to get
-   * @return double array value of the entry
-   */
-  public static double[] getLimelightNetworkTableDoubleArray(String tableName, String entryName) {
-    return getLimelightNetworkTableEntry(tableName, entryName).getDoubleArray(new double[0]);
-  }
-
   /**
    * Gets if the Limelight have a valid target?
    *
@@ -341,7 +29,7 @@ public class TigerHelpers {
    * @return True if a valid target is present, false otherwise
    */
   public static boolean getTV(String limelightName) {
-    return 1.0 == getLimelightNetworkTableDouble(limelightName, "tv");
+    return 1.0 == NTUtils.getLimelightNetworkTableDouble(limelightName, "tv");
   }
 
   /**
@@ -351,7 +39,7 @@ public class TigerHelpers {
    * @return Horizontal offset angle in degrees
    */
   public static double getTX(String limelightName) {
-    return getLimelightNetworkTableDouble(limelightName, "tx");
+    return NTUtils.getLimelightNetworkTableDouble(limelightName, "tx");
   }
 
   /**
@@ -361,7 +49,7 @@ public class TigerHelpers {
    * @return Vertical offset angle in degrees
    */
   public static double getTY(String limelightName) {
-    return getLimelightNetworkTableDouble(limelightName, "ty");
+    return NTUtils.getLimelightNetworkTableDouble(limelightName, "ty");
   }
 
   /**
@@ -373,7 +61,7 @@ public class TigerHelpers {
    * @return Horizontal offset angle in degrees
    */
   public static double getTXNC(String limelightName) {
-    return getLimelightNetworkTableDouble(limelightName, "txnc");
+    return NTUtils.getLimelightNetworkTableDouble(limelightName, "txnc");
   }
 
   /**
@@ -385,7 +73,7 @@ public class TigerHelpers {
    * @return Vertical offset angle in degrees
    */
   public static double getTYNC(String limelightName) {
-    return getLimelightNetworkTableDouble(limelightName, "tync");
+    return NTUtils.getLimelightNetworkTableDouble(limelightName, "tync");
   }
 
   /**
@@ -395,7 +83,7 @@ public class TigerHelpers {
    * @return Target area percentage (0-100%)
    */
   public static double getTA(String limelightName) {
-    return getLimelightNetworkTableDouble(limelightName, "ta");
+    return NTUtils.getLimelightNetworkTableDouble(limelightName, "ta");
   }
 
   /**
@@ -405,7 +93,7 @@ public class TigerHelpers {
    * @return Pipeline latency in milliseconds
    */
   public static double getLatencyPipeline(String limelightName) {
-    return getLimelightNetworkTableDouble(limelightName, "tl");
+    return NTUtils.getLimelightNetworkTableDouble(limelightName, "tl");
   }
 
   /**
@@ -415,21 +103,54 @@ public class TigerHelpers {
    * @return Capture latency in milliseconds
    */
   public static double getLatencyCapture(String limelightName) {
-    return getLimelightNetworkTableDouble(limelightName, "cl");
+    return NTUtils.getLimelightNetworkTableDouble(limelightName, "cl");
+  }
+
+  /**
+   * Gets the latest raw fiducial/AprilTag detection results from NetworkTables.
+   *
+   * @param limelightName The name of the Limelight set in the UI ("" for default)
+   * @return Array of RawFiducial objects containing detection details
+   */
+  public static RawFiducial[] getRawFiducials(String limelightName) {
+    NetworkTableEntry entry = NTUtils.getLimelightNetworkTableEntry(limelightName, "rawfiducials");
+    double[] rawFiducialArray = entry.getDoubleArray(new double[0]);
+    int valsPerEntry = 7;
+    if (rawFiducialArray.length % valsPerEntry != 0) {
+      return new RawFiducial[0];
+    }
+
+    int numFiducials = rawFiducialArray.length / valsPerEntry;
+    RawFiducial[] rawFiducials = new RawFiducial[numFiducials];
+
+    for (int i = 0; i < numFiducials; i++) {
+      int baseIndex = i * valsPerEntry;
+      int id = (int) DataUtils.extractArrayEntry(rawFiducialArray, baseIndex);
+      double txnc = DataUtils.extractArrayEntry(rawFiducialArray, baseIndex + 1);
+      double tync = DataUtils.extractArrayEntry(rawFiducialArray, baseIndex + 2);
+      double ta = DataUtils.extractArrayEntry(rawFiducialArray, baseIndex + 3);
+      double distToCamera = DataUtils.extractArrayEntry(rawFiducialArray, baseIndex + 4);
+      double distToRobot = DataUtils.extractArrayEntry(rawFiducialArray, baseIndex + 5);
+      double ambiguity = DataUtils.extractArrayEntry(rawFiducialArray, baseIndex + 6);
+
+      rawFiducials[i] = new RawFiducial(id, txnc, tync, ta, distToCamera, distToRobot, ambiguity);
+    }
+
+    return rawFiducials;
   }
 
   // TODO: deprecate these methods (all the way down to the enum)
   public static double[] getBotPose(String limelightName) {
-    return getLimelightNetworkTableDoubleArray(limelightName, "botpose");
+    return NTUtils.getLimelightNetworkTableDoubleArray(limelightName, "botpose");
   }
 
   public static double getFiducialID(String limelightName) {
-    return getLimelightNetworkTableDouble(limelightName, "tid");
+    return NTUtils.getLimelightNetworkTableDouble(limelightName, "tid");
   }
 
   public static Pose3d getBotPose3d(String limelightName) {
-    double[] poseArray = getLimelightNetworkTableDoubleArray(limelightName, "botpose");
-    return toPose3D(poseArray);
+    double[] poseArray = NTUtils.getLimelightNetworkTableDoubleArray(limelightName, "botpose");
+    return DataUtils.toPose3D(poseArray);
   }
 
   /**
@@ -441,8 +162,9 @@ public class TigerHelpers {
    */
   @Deprecated
   public static Pose3d getBotPose3d_wpiRed(String limelightName) {
-    double[] poseArray = getLimelightNetworkTableDoubleArray(limelightName, "botpose_wpired");
-    return toPose3D(poseArray);
+    double[] poseArray =
+        NTUtils.getLimelightNetworkTableDoubleArray(limelightName, "botpose_wpired");
+    return DataUtils.toPose3D(poseArray);
   }
 
   /**
@@ -454,8 +176,9 @@ public class TigerHelpers {
    */
   @Deprecated
   public static Pose3d getBotPose3d_wpiBlue(String limelightName) {
-    double[] poseArray = getLimelightNetworkTableDoubleArray(limelightName, "botpose_wpiblue");
-    return toPose3D(poseArray);
+    double[] poseArray =
+        NTUtils.getLimelightNetworkTableDoubleArray(limelightName, "botpose_wpiblue");
+    return DataUtils.toPose3D(poseArray);
   }
 
   /**
@@ -466,8 +189,9 @@ public class TigerHelpers {
    */
   @Deprecated
   public static Pose3d getBotPose3d_TargetSpace(String limelightName) {
-    double[] poseArray = getLimelightNetworkTableDoubleArray(limelightName, "botpose_targetspace");
-    return toPose3D(poseArray);
+    double[] poseArray =
+        NTUtils.getLimelightNetworkTableDoubleArray(limelightName, "botpose_targetspace");
+    return DataUtils.toPose3D(poseArray);
   }
 
   /**
@@ -479,8 +203,8 @@ public class TigerHelpers {
   @Deprecated
   public static Pose3d getCameraPose3d_TargetSpace(String limelightName) {
     double[] poseArray =
-        getLimelightNetworkTableDoubleArray(limelightName, "camerapose_targetspace");
-    return toPose3D(poseArray);
+        NTUtils.getLimelightNetworkTableDoubleArray(limelightName, "camerapose_targetspace");
+    return DataUtils.toPose3D(poseArray);
   }
 
   /**
@@ -492,8 +216,8 @@ public class TigerHelpers {
   @Deprecated
   public static Pose3d getTargetPose3d_CameraSpace(String limelightName) {
     double[] poseArray =
-        getLimelightNetworkTableDoubleArray(limelightName, "targetpose_cameraspace");
-    return toPose3D(poseArray);
+        NTUtils.getLimelightNetworkTableDoubleArray(limelightName, "targetpose_cameraspace");
+    return DataUtils.toPose3D(poseArray);
   }
 
   /**
@@ -505,8 +229,8 @@ public class TigerHelpers {
   @Deprecated
   public static Pose3d getTargetPose3d_RobotSpace(String limelightName) {
     double[] poseArray =
-        getLimelightNetworkTableDoubleArray(limelightName, "targetpose_robotspace");
-    return toPose3D(poseArray);
+        NTUtils.getLimelightNetworkTableDoubleArray(limelightName, "targetpose_robotspace");
+    return DataUtils.toPose3D(poseArray);
   }
 
   /**
@@ -518,67 +242,8 @@ public class TigerHelpers {
   @Deprecated
   public static Pose3d getCameraPose3d_RobotSpace(String limelightName) {
     double[] poseArray =
-        getLimelightNetworkTableDoubleArray(limelightName, "camerapose_robotspace");
-    return toPose3D(poseArray);
-  }
-
-  /**
-   * This enum represents the different types of botpose data that can be retrieved from the
-   * Limelight. Each constant corresponds to a specific NetworkTables entry name and indicates
-   * whether it uses the MegaTag2 algorithm or the original MegaTag1 algorithm for pose estimation.
-   */
-  public enum Botpose {
-    /**
-     * Botpose data for the blue alliance using the MegaTag1 algorithm. Retrieved from the
-     * "botpose_wpiblue" NetworkTables entry.
-     */
-    BLUE_MEGATAG1("botpose_wpiblue", false),
-
-    /**
-     * Botpose data for the blue alliance using the MegaTag2 algorithm. Retrieved from the
-     * "botpose_orb_wpiblue" NetworkTables entry.
-     */
-    BLUE_MEGATAG2("botpose_orb_wpiblue", true),
-
-    /**
-     * Botpose data for the red alliance using the MegaTag1 algorithm. Retrieved from the
-     * "botpose_wpired" NetworkTables entry.
-     */
-    RED_MEGATAG1("botpose_wpired", false),
-
-    /**
-     * Botpose data for the red alliance using the MegaTag2 algorithm. Retrieved from the
-     * "botpose_orb_wpired" NetworkTables entry.
-     */
-    RED_MEGATAG2("botpose_orb_wpired", true);
-
-    private final String entryName;
-
-    private final boolean isMegaTag2;
-
-    Botpose(String entryName, boolean isMegaTag2) {
-      this.entryName = entryName;
-      this.isMegaTag2 = isMegaTag2;
-    }
-
-    /**
-     * Gets the entry name for the botpose type. This is the name of the network table entry that
-     * contains the botpose data.
-     *
-     * @return The network table entry name
-     */
-    public String getEntryName() {
-      return entryName;
-    }
-
-    /**
-     * Gets if the botpose is calculated using MegaTag2.
-     *
-     * @return True if the botpose is calculated using MegaTag2, false if using MegaTag1
-     */
-    public boolean isMegaTag2() {
-      return isMegaTag2;
-    }
+        NTUtils.getLimelightNetworkTableDoubleArray(limelightName, "camerapose_robotspace");
+    return DataUtils.toPose3D(poseArray);
   }
 
   /**
@@ -590,7 +255,7 @@ public class TigerHelpers {
    */
   @Deprecated
   public static PoseEstimate getBotPoseEstimate_wpiBlue(String limelightName) {
-    return unpackBotPoseEstimate(limelightName, "botpose_wpiblue", false);
+    return DataUtils.unpackBotPoseEstimate(limelightName, "botpose_wpiblue", false);
   }
 
   /**
@@ -603,7 +268,7 @@ public class TigerHelpers {
    */
   @Deprecated
   public static PoseEstimate getBotPoseEstimate_wpiBlue_MegaTag2(String limelightName) {
-    return unpackBotPoseEstimate(limelightName, "botpose_orb_wpiblue", true);
+    return DataUtils.unpackBotPoseEstimate(limelightName, "botpose_orb_wpiblue", true);
   }
 
   /**
@@ -615,7 +280,7 @@ public class TigerHelpers {
    */
   @Deprecated
   public static PoseEstimate getBotPoseEstimate_wpiRed(String limelightName) {
-    return unpackBotPoseEstimate(limelightName, "botpose_wpired", false);
+    return DataUtils.unpackBotPoseEstimate(limelightName, "botpose_wpired", false);
   }
 
   /**
@@ -627,7 +292,7 @@ public class TigerHelpers {
    */
   @Deprecated
   public static PoseEstimate getBotPoseEstimate_wpiRed_MegaTag2(String limelightName) {
-    return unpackBotPoseEstimate(limelightName, "botpose_orb_wpired", true);
+    return DataUtils.unpackBotPoseEstimate(limelightName, "botpose_orb_wpired", true);
   }
 
   /**
@@ -639,8 +304,9 @@ public class TigerHelpers {
    * @return the Pose2d of the robot relative to the origin specified by the botpose
    */
   public static Pose2d getBotPose2d(String limelightName, Botpose botpose) {
-    double[] result = getLimelightNetworkTableDoubleArray(limelightName, botpose.getEntryName());
-    return toPose2D(result);
+    double[] result =
+        NTUtils.getLimelightNetworkTableDoubleArray(limelightName, botpose.getEntryName());
+    return DataUtils.toPose2D(result);
   }
 
   /**
@@ -662,7 +328,8 @@ public class TigerHelpers {
    * @return the PoseEstimate object
    */
   public static PoseEstimate getBotPoseEstimate(String limelightName, Botpose botpose) {
-    return unpackBotPoseEstimate(limelightName, botpose.getEntryName(), botpose.isMegaTag2());
+    return DataUtils.unpackBotPoseEstimate(
+        limelightName, botpose.getEntryName(), botpose.isMegaTag2());
   }
 
   /**
@@ -692,7 +359,7 @@ public class TigerHelpers {
     int fiducialCount = poseEstimate.rawFiducials.length;
     double[] data = new double[11 + 7 * fiducialCount];
 
-    // Populates the PoseEstimate, matching unpackBotPoseEstimate
+    // Populates the PoseEstimate, matching DataUtils.unpackBotPoseEstimate
     data[0] = poseEstimate.pose.getX(); // x
     data[1] = poseEstimate.pose.getY(); // y
     data[2] = 0.0; // z (Pose2d doesn't use this)
@@ -719,7 +386,7 @@ public class TigerHelpers {
     }
 
     // Write the array to NetworkTables
-    setLimelightNetworkTableDoubleArray(limelightName, botpose.getEntryName(), data);
+    NTUtils.setLimelightNetworkTableDoubleArray(limelightName, botpose.getEntryName(), data);
   }
 
   /**
@@ -755,7 +422,7 @@ public class TigerHelpers {
    * @param limelightName The name of the Limelight set in the UI ("" for default)
    */
   public static void setRawFiducials(RawFiducial[] rawFiducials, String limelightName) {
-    NetworkTableEntry entry = getLimelightNetworkTableEntry(limelightName, "rawfiducials");
+    NetworkTableEntry entry = NTUtils.getLimelightNetworkTableEntry(limelightName, "rawfiducials");
     double[] data = new double[rawFiducials.length * 7];
 
     for (int i = 0; i < rawFiducials.length; i++) {
@@ -782,7 +449,7 @@ public class TigerHelpers {
    * @return IMUData object containing all current IMU data
    */
   public static IMUData getIMUData(String limelightName) {
-    double[] imuData = getLimelightNetworkTableDoubleArray(limelightName, "imu");
+    double[] imuData = NTUtils.getLimelightNetworkTableDoubleArray(limelightName, "imu");
     if (imuData == null || imuData.length < 10) {
       return new IMUData(); // Returns object with all zeros
     }
@@ -807,7 +474,7 @@ public class TigerHelpers {
    * @param ID ID of the tag to prioritize
    */
   public static void setPriorityTagID(String limelightName, int ID) {
-    setLimelightNetworkTableDouble(limelightName, "priorityid", ID);
+    NTUtils.setLimelightNetworkTableDouble(limelightName, "priorityid", ID);
   }
 
   /**
@@ -828,7 +495,7 @@ public class TigerHelpers {
     entries[1] = cropXMax;
     entries[2] = cropYMin;
     entries[3] = cropYMax;
-    setLimelightNetworkTableDoubleArray(limelightName, "crop", entries);
+    NTUtils.setLimelightNetworkTableDoubleArray(limelightName, "crop", entries);
   }
 
   /**
@@ -845,7 +512,7 @@ public class TigerHelpers {
     entries[0] = offsetX;
     entries[1] = offsetY;
     entries[2] = offsetZ;
-    setLimelightNetworkTableDoubleArray(limelightName, "fiducial_offset_set", entries);
+    NTUtils.setLimelightNetworkTableDoubleArray(limelightName, "fiducial_offset_set", entries);
   }
 
   /**
@@ -874,8 +541,8 @@ public class TigerHelpers {
     entries[3] = pitchRate;
     entries[4] = roll;
     entries[5] = rollRate;
-    setLimelightNetworkTableDoubleArray(limelightName, "robot_orientation_set", entries);
-    flushNetworkTable();
+    NTUtils.setLimelightNetworkTableDoubleArray(limelightName, "robot_orientation_set", entries);
+    NTUtils.flushNetworkTable();
   }
 
   /**
@@ -962,7 +629,7 @@ public class TigerHelpers {
    */
   @Deprecated
   public static void setIMUMode(String limelightName, int imuMode) {
-    setLimelightNetworkTableDouble(limelightName, "imumode_set", imuMode);
+    NTUtils.setLimelightNetworkTableDouble(limelightName, "imumode_set", imuMode);
   }
 
   /**
@@ -972,7 +639,7 @@ public class TigerHelpers {
    * @param imuMode The IMU mode to set, uses the {@link IMUMode} enum.
    */
   public static void setIMUMode(String limelightName, IMUMode imuMode) {
-    setLimelightNetworkTableDouble(limelightName, "imumode_set", imuMode.getModeValue());
+    NTUtils.setLimelightNetworkTableDouble(limelightName, "imumode_set", imuMode.getModeValue());
   }
 
   /**
@@ -984,7 +651,7 @@ public class TigerHelpers {
    *     1 = process every other frame, etc.
    */
   public static void setLimelightThrottle(String limelightName, int throttle) {
-    setLimelightNetworkTableDouble(limelightName, "throttle_set", throttle);
+    NTUtils.setLimelightNetworkTableDouble(limelightName, "throttle_set", throttle);
   }
 
   /**
@@ -1001,7 +668,7 @@ public class TigerHelpers {
     entries[0] = x;
     entries[1] = y;
     entries[2] = z;
-    setLimelightNetworkTableDoubleArray(limelightName, "fiducial_offset_set", entries);
+    NTUtils.setLimelightNetworkTableDoubleArray(limelightName, "fiducial_offset_set", entries);
   }
 
   /**
@@ -1016,7 +683,8 @@ public class TigerHelpers {
     for (int i = 0; i < validIDs.length; i++) {
       validIDsDouble[i] = validIDs[i];
     }
-    setLimelightNetworkTableDoubleArray(limelightName, "fiducial_id_filters_set", validIDsDouble);
+    NTUtils.setLimelightNetworkTableDoubleArray(
+        limelightName, "fiducial_id_filters_set", validIDsDouble);
   }
 
   /**
@@ -1044,7 +712,7 @@ public class TigerHelpers {
     if (downscale == 4) {
       d = 5;
     }
-    setLimelightNetworkTableDouble(limelightName, "fiducial_downscale_set", d);
+    NTUtils.setLimelightNetworkTableDouble(limelightName, "fiducial_downscale_set", d);
   }
 
   /**
@@ -1073,6 +741,7 @@ public class TigerHelpers {
     entries[3] = roll;
     entries[4] = pitch;
     entries[5] = yaw;
-    setLimelightNetworkTableDoubleArray(limelightName, "camerapose_robotspace_set", entries);
+    NTUtils.setLimelightNetworkTableDoubleArray(
+        limelightName, "camerapose_robotspace_set", entries);
   }
 }
